+++ SCE-WebPortal/src/main/java/cn/com/bonc/sce/controller/NavigationController.java
@@ -61,7 +61,7 @@ public class NavigationController {
     @ApiOperation(value = "获取学校机构对应的banner", notes="根据学校id，返回学校对应banner列表", httpMethod = "GET")
     @GetMapping("/getBanners")
     @ResponseBody
-    public RestRecord getBanners(@RequestParam ( "schoolId" ) Integer schoolId ){
+    public RestRecord getBanners(@RequestParam ( "schoolId" ) long schoolId ){
         return navigationService.getBanners(schoolId);
     }
 
@@ -69,7 +69,7 @@ public class NavigationController {
     @PutMapping("/editDefaultBanner")
     @ResponseBody
     public RestRecord editDefaultBanner(
-            @RequestParam ( "schoolId" ) Integer schoolId,
+            @RequestParam ( "schoolId" ) long schoolId,
             @RequestParam( "defaultBanner" ) Integer defaultBanner){
         return navigationService.editDefaultBanner(schoolId,defaultBanner);
     }
@@ -78,7 +78,7 @@ public class NavigationController {
     @DeleteMapping("/delBanner")
     @ResponseBody
     public RestRecord delBanner(
-            @RequestParam( "bannerId" ) Integer bannerId){
+            @RequestParam( "bannerId" ) long bannerId){
         return navigationService.delBanner(bannerId);
     }
 }
